Factor out SliderMarkers
